Move copy/paste code into it's own file
