+++ packages/codemirror-blocks/src/ui/PrimitiveList.tsx
@@ -6,13 +6,13 @@ import {
 } from "../parsers/primitives";
 import { ItemTypes } from "../dnd";
 import { say } from "../announcer";
-import { copy } from "../state/actions";
 import * as selectors from "../state/selectors";
 import CodeMirror from "codemirror";
 import { defaultKeyMap } from "../keymap";
 import { useSelector } from "react-redux";
 import { RootState } from "../state/reducers";
 import { useDrag } from "react-dnd";
+import { copy } from "../copypaste";
 
 require("./PrimitiveList.less");
